fix issue regarding incorrect api call in performtest Slice retreive data, show loading until topic data res received in select catergory component
