+++ components/PerformTest_components/layout/SelectCatergory/SubCatergory.component.jsx
@@ -2,15 +2,23 @@ import React, { useEffect } from "react";
 import { Fragment } from "react";
 import SubCatergoryContent from "./SubCatergoryContent.component";
 import { useSelector, useDispatch } from "react-redux";
+import { retrieveTopicNames } from "../../../../store/performTestSlice";
 
 function SubCatergory() {
   const dispatch = useDispatch();
   const isAlphabetChecked = useSelector(
     (state) => state.performtest.isAlphabetChecked,
   );
+  const storeGitStatus = useSelector(
+    (state) => state.adminDashboard.storeGitStatus,
+  );
   const isRandomChecked = useSelector(
     (state) => state.performtest.isRandomChecked,
   );
+
+  useEffect(() => {
+    dispatch(retrieveTopicNames());
+  }, [storeGitStatus]);
 
   const listOfAlphabets = useSelector((state) =>
     state.performtest.listOfAlphabets.toUpperCase().split(""),
