+++ pages/performtest/index.js
@@ -22,29 +22,20 @@ import Dashboard from "../../components/PerformTest_components/layout/Dashboard/
 import PerformTestHome_Btn from "../../components/PerformTest_components/layout/PerformTestHome_btn";
 import useHydrateState from "../../components/useHydrateState";
 import AdminToolbar from "../../components/Common_Layout/AdminToolbar/AdminToolbar";
-import { retrieveTopicNames } from "../../../../store/performTestSlice";
 
 function index() {
-  const dispatch = useDispatch();
   const showStartBtn = useSelector((state) => state.performtest.showStartBtn);
   const classes = performtestStyles();
   const [hydrateWithLocalStorage] = useHydrateState();
   const showInstructions = useSelector(
     (state) => state.performtest.showInstructions,
   );
-  const storeGitStatus = useSelector(
-    (state) => state.adminDashboard.storeGitStatus,
-  );
+
   const loading = useSelector((state) => state.user.loading);
 
   const showCatergoryForm = useSelector(
     (state) => state.performtest.showCatergoryForm,
   );
-
-  useEffect(() => {
-    if (storeGitStatus !== "success") return;
-    dispatch(retrieveTopicNames());
-  }, [storeGitStatus]);
 
   const showCountdown = useSelector((state) => state.performtest.showCountdown);
   const isTimeIsUp = useSelector((state) => state.performtest.isTimeIsUp);
